% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ennustedatat_demo.R
\name{data_ntp_vuosisumma}
\alias{data_ntp_vuosisumma}
\title{Neljännesvuositilinpito: Bruttokansantuote ja -tulo sekä tarjonta ja kysyntä}
\usage{
data_ntp_vuosisumma()
}
\description{
https://pxnet2.stat.fi/PXWeb/pxweb/fi/StatFin/StatFin__kan__ntp/statfin_ntp_pxt_132h.px/
}
\details{
Muunnos: vuosisumma

Tiedot:
Alkuperäinen sarja käypiin hintoihin, miljoonaa euroa
Alkuperäinen sarja, viitevuosi 2015, miljoonaa euroa
}
