% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ennustedatat_demo.R
\name{data_tyti_vuosikeskiarvot}
\alias{data_tyti_vuosikeskiarvot}
\title{Työvoimatutkimus}
\usage{
data_tyti_vuosikeskiarvot()
}
\description{
https://pxnet2.stat.fi/PXWeb/pxweb/fi/StatFin/StatFin__tym__tyti__kk/statfin_tyti_pxt_135y.px/
}
\details{
Muunnos: vuosikeskiarvo

Sukupuoli:
Yhteensä

Ikäluokka:
15-74
15-64
}
