% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ennustedatat_demo.R
\name{data_vtp_bkt}
\alias{data_vtp_bkt}
\title{Vuositilinpito: Bruttokansantuote ja -tulo sekä tarjonta ja kysyntä
https://pxnet2.stat.fi/PXWeb/pxweb/fi/StatFin/StatFin__kan__vtp/statfin_vtp_pxt_11sf.px/}
\usage{
data_vtp_bkt()
}
\description{
Tiedot:
Käypiin hintoihin, miljoonaa euroa
Edellisen vuoden hinnoin, miljoonaa euroa
}
