% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{draw_ennuste}
\alias{draw_ennuste}
\title{Draw using yaml specification and corresponding prediction data from excel}
\usage{
draw_ennuste(yaml_path, yaml_kuvion_nimi, excel_path, from_year = NULL)
}
\arguments{
\item{path}{path to excel file containing prediction data}
}
\value{
plotly object
}
\description{
Draw using yaml specification and corresponding prediction data from excel
}
\examples{
\dontrun{
piirtaja$using_yaml_and_ennuste_excel(yaml_path = system.file("ennustekuvat/test_without_ennusteet.yaml", package="pttrobo"),
 excel_path = "ptt_ennusteet_KT_uusi_sarja_id.xlsx")
}
}
