% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ennuste_time_serie_from_excel}
\alias{ennuste_time_serie_from_excel}
\title{Read ennustedata from excel file and put it into plotly friendly time serie}
\usage{
ennuste_time_serie_from_excel(excel_path, serie_name)
}
\arguments{
\item{excel_path}{Path to excel file}

\item{serie_name}{name of the serie/identifier}
}
\value{
list object containing last two predictions for the serie
}
\description{
Read ennustedata from excel file and put it into plotly friendly time serie
}
\examples{
\dontrun{
ennusteet_from_excel(excel_path = "ptt_ennusteet_KT.xlsx",
                   serie_name = "StatFin/kan/ntp/statfin_ntp_pxt_132h.px§B1GMH§kausitvv2015")
}
}
