% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_data_yaml.R
\name{excel_files_ennustedata}
\alias{excel_files_ennustedata}
\title{Get excel files from ennustedata folder}
\usage{
excel_files_ennustedata(pattern = NULL)
}
\arguments{
\item{pattern}{A string to look in excel file names.}
}
\description{
Get excel files from ennustedata folder
}
\examples{
excel_files_ennustedata()
excel_files_ennustedata("ME")
}
