% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{excel_to_tibble}
\alias{excel_to_tibble}
\title{Parse PTT prediction data from excel to tibble format}
\usage{
excel_to_tibble(excel_path)
}
\arguments{
\item{excel_path}{Path to the excel file containing prediction data}
}
\value{
prediction data in tibble format
}
\description{
Parse PTT prediction data from excel to tibble format
}
\examples{
\dontrun{
excel_to_tibble("ennustedata.xlsx")
}
}
