% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ennustedatat_demo.R
\name{ntp_tunnit}
\alias{ntp_tunnit}
\title{Neljännesvuositilinpito: Työllisyys ja työtunnit}
\usage{
ntp_tunnit()
}
\description{
https://pxnet2.stat.fi/PXWeb/pxweb/fi/StatFin/StatFin__kan__ntp/statfin_ntp_pxt_11tj.px/
}
\details{
Muunnos: vuosisumma (tunnit), vuosikeskiarvo (muut)

Tiedot:
Alkuperäinen sarja
}
