% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_data_yaml.R
\name{ptt_copy_ennustedata}
\alias{ptt_copy_ennustedata}
\title{Copy ennustedata to teams}
\usage{
ptt_copy_ennustedata(ennuste = c("ME", "KT", "MA", "POP"), path = NULL)
}
\arguments{
\item{ennuste}{A character of ennuste. One of "ME", "KT", "MA".}
}
\description{
Copy ennustedata to teams
}
