% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ptt_plot}
\alias{ptt_plot}
\title{Plot data with plotly}
\usage{
ptt_plot()
}
\arguments{
\item{data_tibble}{data for plotting in tibble format}
}
\value{
plotly object
}
\description{
Plot data with plotly
}
\examples{
\dontrun{
ptt_plot()$lines(d,
                title = "otsikko",
                lahde = "lähde",
                yksikko = "yksikkö")
}
}
