% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptt_plot_base.R
\name{ptt_plot_add_prediction}
\alias{ptt_plot_add_prediction}
\title{Add prediction traces to an existing ptt_plot.}
\usage{
ptt_plot_add_prediction(
  p,
  pred_data,
  grouping = sarja_nmi,
  n_obs = 2,
  with_labs = T,
  isolate_primary = F,
  showlegend = F,
  hovertext = list(rounding = 1, unit = "", extra = "(ennuste)", dateformat = "Annual"),
  value_multiplier = 1,
  custom_pred_data = FALSE
)
}
\arguments{
\item{p}{Plotly object created with ptt_plot to add prediction traces to.}

\item{pred_data}{Tibble in ptt prediction data format.}

\item{grouping}{Tibble column used for grouping in plot, should be labeled the same as data used for parent ptt_plot.}

\item{n_obs}{Number of observations (counted from the latest) used from the prediction set pred_data.}

\item{isolate_primary}{Separates the first factor in parent plot grouping by line width.}

\item{showlegend, with_labs}{Controls prediction trace legend. Hovertemplate takes labeling based on with_labs.}

\item{hovertext}{A list describing hovertext items "list(rounding = 1, unit = "\%", extra = "(ennuste)")".}

\item{value_multiplier}{A number. Value of 0.001 would cause values to be divided by 1000, value of 1000 would cause values to be multiplied by 1000 hovertext items "list(rounding = 1, unit = "\%", extra = "(ennuste)")".}

\item{custom_pred_data}{A boolean. If set to true, pred_data is required to be a dataframe/tibble with columns, year, sarja_nmi and value. With two prediction values per sarja_nmi.}
}
\value{
plotly object
}
\description{
Outputs a plotly object.
}
\examples{
e <- readxl::read_excel("ptt_ennusteet_KT.xlsx") |> dplyr::filter(stringr::str_detect(filter, "B1GMH|P3KS14"))
p <- p |> ptt_plot_add_prediction(e)
p
}
