% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptt_plot_base.R
\name{ptt_plot_add_prediction_traces}
\alias{ptt_plot_add_prediction_traces}
\title{Add prediction traces to an existing ptt_plot.}
\usage{
ptt_plot_add_prediction_traces(
  p,
  pred_data,
  grouping = sarja_nmi,
  n_obs = 2,
  with_labs = T,
  isolate_primary = F,
  showlegend = F,
  hovertext = list(rounding = 1, unit = "\%", extra = "(ennuste)")
)
}
\arguments{
\item{p}{Plotly object created with ptt_plot to add prediction traces to.}

\item{pred_data}{Tibble in ptt prediction data format.}

\item{grouping}{Tibble column used for grouping in plot, should be labeled the same as data used for parent ptt_plot.}

\item{n_obs}{Number of observations (counted from the latest) used from the prediction set pred_data.}

\item{isolate_primary}{Separates the first factor in parent plot grouping by line width.}

\item{showlegend, with_labs}{Controls prediction trace legend. Hovertemplate takes labeling based on with_labs.}

\item{hovertext}{A list describing hovertext items "list(rounding = 1, unit = "\%", extra = "(ennuste)")".}
}
\value{
plotly object
}
\description{
Outputs a plotly object.
}
\examples{
e <- readxl::read_excel("ptt_ennusteet_KT.xlsx") |> dplyr::filter(stringr::str_detect(sarja, "B1GMH|P3KS14"))
p <- p |> ptt_plot_add_prediction_traces(e)
p
}
