% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptt_plot_base.R
\name{ptt_plot_add_secondary_traces}
\alias{ptt_plot_add_secondary_traces}
\title{Add secondary traces of existing traces with reduced linewidth to an existing ptt_plot.}
\usage{
ptt_plot_add_secondary_traces(
  p,
  secondary_data,
  relates_to,
  grouping,
  hovertext,
  showlegend = TRUE
)
}
\arguments{
\item{p}{Plotly object created with ptt_plot to add prediction traces to.}

\item{secondary_data}{Tibble with secondary variables of one grouping in the parent ptt_plot object.}

\item{relates_to}{The name of the grouping that the secondary data relates.}

\item{grouping}{Tibble column used for grouping in plot, should be labeled the same as data used for parent ptt_plot.}

\item{hovertext}{Uses parent ptt_plot specification if undefined. A list describing hovertext items "list(rounding = 1, unit = "\%", extra = "(ennuste)")".}

\item{showlegend}{A locigal to show legend for secondary trace.}
}
\value{
plotly object
}
\description{
Outputs a plotly object.
}
\examples{
d2 <- d |> dplyr::filter(tiedot == "BKT") |> dplyr::mutate(value = statfitools::trend_series(value, time)) |> dplyr::mutate(tiedot = "esimerkkisarja")
}
