% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptt_plot_extra.R
\name{ptt_plot_add_secondary_traces2}
\alias{ptt_plot_add_secondary_traces2}
\title{Add secondary traces of existing traces with reduced linewidth to an existing ptt_plot.}
\usage{
ptt_plot_add_secondary_traces2(
  p,
  secondary_data,
  y = value,
  grouping,
  name = NULL,
  line_width = 2,
  hovertext,
  showlegend = FALSE,
  grouplegend = TRUE
)
}
\arguments{
\item{p}{Plotly object created with ptt_plot to add prediction traces to.}

\item{secondary_data}{Tibble with secondary variables of one grouping in the parent ptt_plot object.}

\item{grouping}{Tibble column used for grouping in plot, should be labeled the same as data used for parent ptt_plot.}

\item{name}{The name of the data in the secondary data.}

\item{line_width}{A width of the line.}

\item{hovertext}{Uses parent ptt_plot specification if undefined. A list describing hovertext items "list(rounding = 1, unit = "\%", extra = "(ennuste)", dateformat = "\%Y-\%m-&d")".}

\item{showlegend}{A locigal to show legend for secondary trace.}

\item{grouplegend}{A locigal to tie legend grouping to grouping in p.}
}
\value{
plotly object
}
\description{
Outputs a plotly object.
}
\examples{
library(tidyverse)
library(pttdatahaku)
dat <- ptt_data_robo("StatFin/ntp/statfin_ntp_pxt_132h.px") |>
  filter_recode(
    taloustoimi = c(
      "BKT" = "B1GMH Bruttokansantuote markkinahintaan",
      "Vienti" = "P61K Tavaroiden vienti, menona"),
    tiedot = c(
     kausitasoitettu = "Kausitasoitetun ja työpäiväkorjatun sarjan volyymin muutos vuodentakaisesta, \%",
     trendi = "Trendisarjan volyymin muutos vuodentakaisesta, \%")) |>
 spread(tiedot, value)
dat |>
  rename(value = trendi) |>
  ptt_plot(grouping = taloustoimi,
           title = "BKT ja vienti",
           subtitle = "\%, volyymin vuosimuutos",
           caption = "Lähde: Tilastokeskus, PTT") |>
  ptt_plot_add_secondary_traces2(dat, y = kausitasoitettu, grouping = taloustoimi)

  dat |>
  rename(value = trendi) |>
  ptt_plot(grouping = taloustoimi,
           title = "BKT ja vienti",
           subtitle = "\%, volyymin vuosimuutos, trendi ja kausitasoitettu",
           caption = "Lähde: Tilastokeskus, PTT") |>
  ptt_plot_add_secondary_traces2(dat, y = kausitasoitettu, grouping = taloustoimi,
  name = "kausitasoitettu", showlegend = TRUE, grouplegend = FALSE)
}
