% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ptt_plot_from}
\alias{ptt_plot_from}
\title{Plot PTT prediction data in plotly format from chosen file with specific time series}
\usage{
ptt_plot_from(excel_path, time_series_name)
}
\arguments{
\item{excel_path}{Excel file's path that contains prediction data}

\item{time_series_name}{Name of time series to be plotted}
}
\value{
plotly object
}
\description{
Plot PTT prediction data in plotly format from chosen file with specific time series
}
\examples{
ptt_plot_from("ptt_ennusteet_testi.xlsx", "bkt_private_consumption")
}
