% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptt_plot_base.R
\name{ptt_plot_set_colors}
\alias{ptt_plot_set_colors}
\title{Gives a vector of desired length of colors.}
\usage{
ptt_plot_set_colors(n_unique, color_vector, accessibility_params)
}
\arguments{
\item{n_unique}{How many colors are needed.}

\item{color_vector}{Optional. A vector or colors.}

\item{accessibility_params}{Placeholder for accessibility features.}
}
\value{
plotly object

Vector of colors.
}
\description{
Outputs a vector of colors.
}
