% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_data_yaml.R
\name{ptt_update_data_sharepoint}
\alias{ptt_update_data_sharepoint}
\title{Update PTT data from yaml definitions}
\usage{
ptt_update_data_sharepoint(
  pattern = NULL,
  site,
  drive = "Documents",
  path,
  start_year,
  transpose = FALSE
)
}
\arguments{
\item{pattern}{A string to look in file names to update.}

\item{start_year}{A first year of data}

\item{transpose}{A logical to transpose data for excel.}
}
\description{
Update PTT data from yaml definitions
}
\examples{
  ptt_update_data_sharepoint(
    site = "Ennuste", path = "KT/Taulut KT/Laskentataulukot/KTdata_test",
    start_year = 2013)
}
