% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_data_yaml.R
\name{ptt_update_ennustedata}
\alias{ptt_update_ennustedata}
\title{Update PTT ennustedata from yaml definitions}
\usage{
ptt_update_ennustedata(pattern = NULL, start_year)
}
\arguments{
\item{pattern}{A string to look in file names to update.}

\item{start_year}{A first year of data}
}
\description{
Update PTT ennustedata from yaml definitions
}
