% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{pttrobo_print_code}
\alias{pttrobo_print_code}
\title{Print retrieval code and full filtering for a Tosi table or URL}
\usage{
pttrobo_print_code(x, conc = TRUE)
}
\arguments{
\item{x}{Tosi table identifier or supported source-table URL.}

\item{conc}{A locigal to copy to clipboard.}
}
\description{
Print retrieval code and full filtering for a Tosi table or URL
}
\examples{
  pttrobo_print_code("https://pxweb2.stat.fi/PxWeb/pxweb/fi/StatFin/StatFin__vaerak/statfin_vaerak_pxt_11ra.px/", conc = FALSE)
  pttrobo_print_code("statfin/vaerak/11ra.px", conc = FALSE)

}
