% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_data_yaml.R
\name{yaml_files_ennustedata}
\alias{yaml_files_ennustedata}
\title{Get yaml files from ennustedata folder}
\usage{
yaml_files_ennustedata(pattern = NULL)
}
\arguments{
\item{pattern}{A string to look in yaml file names.}
}
\description{
Get yaml files from ennustedata folder
}
\examples{
yaml_files_ennustedata()
yaml_files_ennustedata("ME")
}
