% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{yaml_to_plotly_data}
\alias{yaml_to_plotly_data}
\title{Read output kuvio yaml file, containing specification for picture. Can handle up to 3 different data spesifications.}
\usage{
yaml_to_plotly_data(file, kuvion_nimi)
}
\arguments{
\item{file}{Path to yaml file}
}
\value{
list object
}
\description{
Output list containing data for plotting and label
}
\examples{
\dontrun{
yaml_to_plotly_data(file = system.file("ennustekuvat/test_without_ennusteet.yaml", package="pttrobo"),
                   kuvion_nimi = "bkt_ja_kulutus")
}
}
