% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{yearly_change}
\alias{yearly_change}
\title{Transforms absolute values to yearly change}
\usage{
yearly_change(data)
}
\arguments{
\item{tibble}{data to be transformed}
}
\value{
tibble
}
\description{
Transforms absolute values to yearly change
}
\examples{
\dontrun{
yearly_change(data = ptt_data_robo("statfin/ntp/132h.px") \%>\%
                     filter(str_detect(taloustoimi, "B1GMH")) \%>\%
                     filter(tiedot \%in\% c("Kausitasoitettu ja työpäiväkorjattu sarja, viitevuosi 2015, miljoonaa euroa")))
}
}
