% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{aplot_lines}
\alias{aplot_lines}
\alias{aplot_trends}
\title{Basic PTT line plot}
\usage{
aplot_lines(
  dat,
  x = time,
  y = value,
  colour = tiedot,
  size = NULL,
  title = "",
  subtitle = "",
  source = NULL,
  caption = "",
  rangeslider = FALSE,
  ...
)

aplot_trends(
  dat,
  x = time,
  y = value,
  colour = tiedot,
  size = NULL,
  title = "",
  subtitle = "",
  source = NULL,
  caption = "",
  rangeslider = FALSE,
  org_showlegend = FALSE,
  trends_only = FALSE,
  ...
)
}
\arguments{
\item{dat}{A data.frame to plot.}

\item{colour}{A variable for colours}

\item{size}{A variable for size.}

\item{title}{A plot title}

\item{subtitle}{A plot subtitle, use also for y-axis label.}

\item{source}{A source information for caption. Text is added to
"Lähde: ..., PTT".}

\item{caption}{A caption text. Will override source}

\item{rangeslider}{A locigal for rangeslider}

\item{...}{Additional arguments for ptt_plot.}

\item{org_showlegend}{A logical to show legend for original data.}

\item{trends_only}{A logical to have only trends in plot.}
}
\description{
Basic PTT line plot
}
\section{Functions}{
\itemize{
\item \code{aplot_trends}: Estimate and plot trend with original
}}

\examples{
ptt_data_robo_l("StatFin/kan/vtp/statfin_vtp_pxt_11sf.px") |>
dplyr::filter(
    taloustoimi \%in\% c("P7R Tavaroiden ja palvelujen tuonti, tulona",
    "P6K Tavaroiden ja palvelujen vienti, menona"),
    tiedot == c("Volyymin muutokset, \%")) |>
    aplot_lines(colour = taloustoimi,
                title = "Vienti ja tuonti",
                subtititle = "\%, volyymin muutos",
                source = "Tilastokeskus")


ptt_data_robo_l("StatFin/kan/ntp/statfin_ntp_pxt_132h.px") |>
dplyr::filter(
    taloustoimi \%in\% c("P7R Tavaroiden ja palvelujen tuonti, tulona",
    "P6K Tavaroiden ja palvelujen vienti, menona"),
    tiedot == c("Alkuperäinen sarja, viitevuosi 2015, miljoonaa euroa")) |>
    aplot_trends(colour = taloustoimi,
                title = "Vienti ja tuonti",
                subtititle = "\%, volyymin muutos",
                source = "Tilastokeskus")
}
