% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{aplot_lines}
\alias{aplot_lines}
\title{Basic PTT line plot}
\usage{
aplot_lines(
  dat,
  x = time,
  y = value,
  colour = tiedot,
  size = NULL,
  title = NULL,
  subtitle = NULL,
  source = NULL,
  caption = NULL,
  ...
)
}
\arguments{
\item{dat}{A data.frame to plot.}

\item{colour}{A variable for colours}

\item{size}{A variable for size.}

\item{title}{A plot title}

\item{subtitle}{A plot subtitle, use also for y-axis label.}

\item{source}{A source information for caption. Text is added to
"Lähde: ..., PTT".}

\item{caption}{A caption text. Will override source}
}
\description{
Basic PTT line plot
}
\examples{
ptt_data_robo_l("StatFin/kan/vtp/statfin_vtp_pxt_11sf.px") |>
dplyr::filter(
    taloustoimi \%in\% c("P7R Tavaroiden ja palvelujen tuonti, tulona",
    "P6K Tavaroiden ja palvelujen vienti, menona"),
    tiedot == c("Volyymin muutokset, \%")) |>
    aplot_lines(colour = taloustoimi,
                title = "Vienti ja tuonti",
                subtititle = "\%, volyymin muutos",
                source = "Tilastokeskus")


}
