% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deflate.R
\name{deflate}
\alias{deflate}
\title{Deflate nominal series to real ones}
\usage{
deflate(
  x,
  time,
  deflator = "eki",
  index = NULL,
  class = NULL,
  freq = "m",
  baseyear = 2015
)
}
\arguments{
\item{x}{A series to deflate}

\item{time}{A time vector.}

\item{deflator}{A name of the deflatior. "eki" or "thi".}

\item{class}{A name of classification for thi}

\item{freq}{A frequency of \code{x}.}

\item{baseyear}{A base year to rebase.}

\item{series}{A name of index series for thi.}
}
\value{
A numeric vector
}
\description{
Gets a price index and aggregates it, if needed. Then
divides the series with the price index and rebase result to base year.
}
\details{
Price index available:
\itemize{
\item eki Elinkustannusindeksi
\item thi Tuottajahintaideksi
}

For thi index and class have be specified:

index, one of: "Teollisuuden tuottajahintaindeksi",
"Teollisuuden tuottajahintaindeksi, kotimaiset tavarat",
"Teollisuuden tuottajahintaindeksi, vientitavarat",
"Vientihintaindeksi",
"Tuontihintaindeksi",
"Kotimarkkinoiden perushintaindeksi",
"Kotimarkkinoiden perushintaindeksi, kotimaiset tavarat",
"Kotimarkkinoiden perushintaindeksi, tuontitavarat",
"Verollinen kotimarkkinoiden perushintaindeksi",
"Verollinen kotimarkkinoiden perushintaindeksi, kotimaiset tavarat",
"Verollinen kotimarkkinoiden perushintaindeksi, tuontitavarat"

class one of Tuotteet toimialoittain (CPA 2015, MIG) classification
}
\examples{
pttrobo::ptt_data_robo("StatFin/ati/statfin_ati_pxt_11zt.px") |>
  filter_recode(
    tiedot = c("Ansiotaso" = "Ansiotasoindeksi 1964=100")
  ) |>
    mutate(real = deflate(value, time, deflator = "eki", freq = "q", baseyear = 2015)) |>
    tail()

 pttrobo::ptt_data_robo(
   "tulli/uljas_sitc",
   dl_filter = list(
     "Tavaraluokitus SITC2" = c("01 (2002--.) Liha ja lihatuotteet"),
     "Maa" = "AA",
     "Suunta" = c("Tuonti alkuperämaittain"),
     "Indikaattorit" = c("Tilastoarvo (euro)"))) |>
     mutate(real = deflate(value, time, deflator = "thi",
                     index = "Tuontihintaindeksi",
                     class = "10.1 Säilötty liha ja lihavalmisteet")) |>
   tail()
}
