% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptt_data.R
\name{ptt_data_robo}
\alias{ptt_data_robo}
\title{Get data with robonomistClient in PTT form}
\usage{
ptt_data_robo(..., labels = FALSE)
}
\arguments{
\item{...}{Arguments to \code{\link[robonomistClient]{data_get}}.
see \code{\link[robonomistClient]{data}}}

\item{labels}{A logical whether to get labels or codes.}
}
\description{
Uses \code{robonomistClient::data_get} to get data from robonomist,
but with PTT formating:
\itemize{
\item codes instead of labels
\item tidy_time = TRUE
\item variables as factors
}
}
\examples{
\dontrun{
  ptt_data_robo("StatFin/asu/asvu/statfin_asvu_pxt_11x4.px") |> head()
}
}
