% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptt_data.R
\name{ptt_data_robo}
\alias{ptt_data_robo}
\alias{ptt_data_robo_l}
\alias{ptt_data_robo_c}
\alias{ptt_data_robo_b}
\title{Get Data with robonomistClient in PTT Format}
\usage{
ptt_data_robo(..., labels = TRUE)

ptt_data_robo_l(..., labels = TRUE)

ptt_data_robo_c(..., labels = FALSE)

ptt_data_robo_b(...)
}
\arguments{
\item{...}{Arguments to \code{\link[robonomistClient]{data_get}}.
see \code{\link[robonomistClient]{data}}}

\item{labels}{A logical whether to get labels or codes.}
}
\description{
Uses \code{robonomistClient::data_get} to get data from robonomist,
but with PTT formating:
\itemize{
\item tidy_time = TRUE
\item variables as factors
}
}
\section{Functions}{
\itemize{
\item \code{ptt_data_robo_l}: With labels TRUE.

\item \code{ptt_data_robo_c}: With labels FALSE.

\item \code{ptt_data_robo_b}: Both labels and codes.
}}

\examples{
\dontrun{
  ptt_data_robo("StatFin/asu/asvu/statfin_asvu_pxt_11x4.px") |> head()
}
}
