% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ptt_plot}
\alias{ptt_plot}
\title{Plot PTT prediction data in plotly format}
\usage{
ptt_plot()
}
\arguments{
\item{data_tibble}{data for plotting in tibble format}
}
\value{
plotly object
}
\description{
Plot PTT prediction data in plotly format
}
\examples{
Filtering datatable that contains multiple different time series
d <- excel_to_tibble("ptt_ennusteet_testi.xlsx") \%>\% filter(name == "bkt_private_consumption")
Using plotting function with parameters
ptt_plot()$lines(d,
                title = d$title,
                lahde = d$caption,
                yksikko = d$ylab)
}
