% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptt_plot_base.R
\name{ptt_plot}
\alias{ptt_plot}
\title{Creates a plotly object with visual specifications of ptt.}
\usage{
ptt_plot(
  d,
  grouping,
  title,
  subtitle = "",
  caption,
  legend_orientation = "auto",
  legend_position = "bottom",
  margin = NA,
  font_color = "#696969",
  grid_color = "#E8E8E8",
  zeroline = F,
  rangeslider = FALSE,
  isolate_primary = F,
  font_size = 16,
  hovertext,
  axis_limits = list(x = c(NA, NA), y = c(NA, NA)),
  plot_type = "scatter",
  height = 600,
  ...
)
}
\arguments{
\item{d}{Data for plotting (Tibble).}

\item{grouping}{Tibble column from tibble d to use for grouping (Unquoted string).}

\item{title, subtitle, caption}{labeling for plot}

\item{legend_orientation, legend_position}{Legend positioning.}

\item{margin}{Plot margins if calculated margins do not work (List).}

\item{font_color, grid_color, font_size}{Plot background element colors and font sizes.}

\item{zeroline}{Determines zeroline inclusion (Logical or double).}

\item{rangeslider}{Determines rangeslider inclusion (Logical).}

\item{isolate_primary}{Separates the first factor in grouping by line width. Use ptt_add_secondary_traces for multiple series with multiple secondary traces.}

\item{hovertext}{A list describing hovertext items "list(rounding = 1, unit = "\%", extra = "(ennuste)")".}

\item{axis_limits}{Determines the limits of the axes (list(x = c(NA,NA), y = c(NA,NA)))".}

\item{plot_type}{Determines the trace type for either the whole plot, or for all variables defined by grouping as name-value pairs (Character vector).}

\item{height}{Height of the plot.}
}
\value{
plotly object
}
\description{
Outputs a plotly object.
}
\examples{
d <- ptt_data_robo_l("StatFin/kan/ntp/statfin_ntp_pxt_132h.px") |>
  dplyr::filter(stringr::str_detect(taloustoimi, "B1GMH|P3KS14_S15"), tiedot == "Kausitasoitettu ja työpäiväkorjattu sarja, viitevuosi 2015, miljoonaa euroa", lubridate::year(time) >= 2010) |>
  dplyr::group_by(taloustoimi) |>
  dplyr::mutate(value = ((value/ lag(value, 4)) -1) * 100) |>
  dplyr::ungroup() |>
  dplyr::mutate(tiedot = dplyr::case_when(stringr::str_detect(taloustoimi, "B1GMH") ~ "BKT", TRUE ~ "Yksityinen kulutus") |> forcats::as_factor()) |>
  tidyr::drop_na()
  p <- d |> ptt_plot(tiedot, "BKT ja kulutus", subtitle = "Vuosimuutos",
  caption =  "Lähde: Tilastokeskus ja PTT",
  rangeslider = "2015-01-01",
  zeroline = T)
  p
}
