% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptt_plot_base.R
\name{ptt_plot}
\alias{ptt_plot}
\title{Creates a plotly object with visual specifications of ptt.}
\usage{
ptt_plot(
  d,
  grouping,
  title,
  subtitle = "",
  caption,
  legend_orientation = "auto",
  legend_position = "bottom",
  margin = NA,
  font_color = "#696969",
  grid_color = "#E8E8E8",
  rangeslider = T,
  secondary_grouping = NA,
  font_size = 14,
  hovertext = list(rounding = 1, unit = "", extra = ""),
  height = 400,
  ...
)
}
\arguments{
\item{d}{Tibble for plotting.}

\item{grouping}{Tibble column from tibble d to use for grouping.}

\item{title, subtitle, caption}{labeling for plot}

\item{legend_orientation, legend_position}{Legend positioning.}

\item{margin}{List for plot margins if calculated margins do not work.}

\item{font_color, grid_color, font_size}{Plot background element colors and font sizes.}

\item{rangeslider}{Determines rangeslider inclusion.}

\item{secondary_grouping}{Additional grouping variable for line widths.}

\item{hovertext}{A list describing hovertext items "list(rounding = 1, unit = "\%", extra = "(ennuste)")".}

\item{height}{Height of the plot.}
}
\value{
plotly object
}
\description{
Outputs a plotly object.
}
\examples{
d <- ptt_data_robo_l("StatFin/kan/ntp/statfin_ntp_pxt_132h.px") |>
dplyr::filter(stringr::str_detect(taloustoimi, "B1GMH|P3KS14_S15"), tiedot == "Kausitasoitettu ja työpäiväkorjattu sarja, viitevuosi 2015, miljoonaa euroa", lubridate::year(time) >= 2010) |>
dplyr::group_by(taloustoimi) |>
dplyr::mutate(value = ((value/ lag(value, 4)) -1) * 100) |>
dplyr::ungroup() |>
dplyr::mutate(tiedot = dplyr::case_when(stringr::str_detect(taloustoimi, "B1GMH") ~ "BKT", TRUE ~ "Yksityinen kulutus") \%>\% forcats::as_factor()
p <- d |> ptt_plot(tiedot, "BKT ja kulutus", subtitle = "Vuosimuutos",
                   caption =  "Lähde: Tilastokeskus ja PTT",lwd = F,rangeslider = T,height = 600)
p
}
