% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptt_plot_widget.R
\name{ptt_plot_create_widget}
\alias{ptt_plot_create_widget}
\title{Writes an html element for embedding, and optionally png files.}
\usage{
ptt_plot_create_widget(
  p,
  title,
  filepath,
  render = T,
  self_contained = F,
  png_artefacts,
  png_folder,
  iframe_height = "550px"
)
}
\arguments{
\item{p}{A plotly object.}

\item{title}{The filename of the html element (without file format). The function will clean the name up, or try to extract it from param p if missing.}

\item{render}{Logical. Is the plot rendered to viewer after saving the widget (default true). Returns the plot object nonetheless.}

\item{self_contained}{Logical. Will the html artefact have self-contained dependencies, increasing size. Default false.}

\item{png_artefacts}{Optional character vector of s(mall), n(arrow), and/or w(ide) corresponding to the expected .png sizes.}

\item{png_folder}{A folder to save png-files.}

\item{path}{The path of the saved file. When knitting and .Rmd file, the a folder is created matching the file name of the currently knit document and the path is set there.}

\item{iframe_heigh}{A height of the iframe.}
}
\value{
The plotly object p.
}
\description{
Writes an html element for embedding, and optionally png files.
}
\examples{
p |> ptt_plot_create_widget()
}
