% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptt_plot_base.R
\name{ptt_plot_create_widget}
\alias{ptt_plot_create_widget}
\title{Writes an html element for embedding.}
\usage{
ptt_plot_create_widget(p, title)
}
\arguments{
\item{p}{A plotly object.}

\item{title}{The filename of the html element (without file format). The function will clean the name up, or try to extract it from param p if missing.}
}
\value{
None
}
\description{
Writes an html element for embedding.
}
\examples{
p |> ptt_plot_create_widget()
}
