% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptt_plot_widget.R
\name{ptt_plot_upload_widgets}
\alias{ptt_plot_upload_widgets}
\title{Uploads the html elements and dependencies to cloud storage.}
\usage{
ptt_plot_upload_widgets(
  files_path,
  upload_path,
  overwrite = FALSE,
  release_time = TRUE
)
}
\arguments{
\item{files_path}{The folder where the artefacts to be uploaded are located.}

\item{upload_path}{The gcs folder where the artefacts will be uploaded to.}

\item{release_time}{A release time in same format as \code{\link[base]{Sys.time}}.
Can be set also \code{FALSE}. If past the Sys.time() of code{TRUE} prevents upload.}
}
\description{
The cloud storage authentication file have to in working folder or in Tiedosto/Documents folder.
}
