% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptt_plot_base.R
\name{ptt_plot_upload_widgets}
\alias{ptt_plot_upload_widgets}
\title{Uploads the html elements and dependencies to cloud storage.}
\usage{
ptt_plot_upload_widgets(files_path, upload_path)
}
\arguments{
\item{files_path}{The folder where the artefacts to be uploaded are located.}

\item{upload_path}{The gcs folder where the artefacts will be uploaded to.}
}
\description{
Uploads the html elements and dependencies to cloud storage.
}
