% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{pttrobo_print_filter}
\alias{pttrobo_print_filter}
\alias{pttrobo_print_filter_recode}
\title{Print full filtering for Robonomist id or data}
\usage{
pttrobo_print_filter(x, conc = TRUE, print = TRUE)

pttrobo_print_filter_recode(x, conc = TRUE, print = TRUE)
}
\arguments{
\item{x}{A Robonomist id or robonomist_data}

\item{conc}{A locigal whether to copy in clipboard}

\item{print}{A locigal whether to print output (to only return invisibly)}
}
\description{
In ptt-format
}
\section{Functions}{
\itemize{
\item \code{pttrobo_print_filter_recode}: version for pttdatahaku::filter_recode()
}}

\examples{
  pttrobo_print_filter(x = "luke/02_Maatalous/06_Talous/02_Maataloustuotteiden_tuottajahinnat/08_Tuottajahinnat_Vilja_rypsi_rapsi_v.px", conc = FALSE)
  pttrobo_print_filter_recode(x = "luke/02_Maatalous/06_Talous/02_Maataloustuotteiden_tuottajahinnat/08_Tuottajahinnat_Vilja_rypsi_rapsi_v.px", conc = FALSE)
}
