% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_yaml.R
\name{yaml_to_excel}
\alias{yaml_to_excel}
\title{Read data yaml file, get data, and output xlxs files}
\usage{
yaml_to_excel(
  file,
  xlsx_path = system.file("ennustedata", package = "pttrobo"),
  start_year
)
}
\arguments{
\item{file}{Path to yaml file}

\item{xlsx_path}{A path to save excel files.}

\item{start_year}{A numeric. Year to start data

#' @export}
}
\description{
Read data yaml file, get data, and output xlxs files
}
