% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{yearly_change}
\alias{yearly_change}
\title{Transforms absolute values to yearly change}
\usage{
yearly_change(data)
}
\arguments{
\item{tibble}{data to be transformed}
}
\value{
tibble
}
\description{
Transforms absolute values to yearly change
}
\examples{
\dontrun{
yearly_change(data = robonomistClient::data("StatFin/kan/ntp/statfin_ntp_pxt_132h.px", tidy_time = TRUE) \%>\%
                     filter(str_detect(Taloustoimi, "B1GMH")) \%>\%
                     filter(Tiedot \%in\% c("Kausitasoitettu ja työpäiväkorjattu sarja, viitevuosi 2015, miljoonaa euroa")))
}
}
